fix(wheel): repair Gem Atelier, Fragment Workshop and presets
Gem actions went through ProtocolGame::sendWheelGemAction with a byte
layout that didn't match the server's playerWheelGemAction parser: the
gem index was dropped for destroy/switch/lock (only action 4 sent a 3rd
byte) and enhance sent fragmentType/pos swapped. Rewrite it per-action
(destroy/switch/lock = U16 index, reveal = U8 quality, improveGrade =
U8 fragmentType + U8 modPos) and widen param to U16; Lua passes the gem
id in param and enhance passes (supreme and 0 or 1, modID).

Grade icons stayed grayscale because setImageShader/getImageShader had
no C++->Lua binding, so widget:setImageShader('') hit the no-op shim in
globals.lua (wrote a Lua field, never m_shader) and the OTUI
image_black_white was never cleared. Bind both (same class as the
UIProgressRect shim bug) so runtime shader toggles work client-wide;
setImageShader now also invalidates the image cache.

Integer-truthiness bugs (Lua treats 0 as true) coerced to == 1: the lock
checkbox setChecked(data.locked), the switch/dismantle gate
(if gemData.locked) and the "Locked only" filter.

Also:
- track workshop selection by mod identity {modID, supreme} instead of
  grid slot so the detail panel no longer desyncs after a re-render
- dim locked grade circles/lines (static green frames hid the state)
- dismantle captures the gem id up front (callback read stale -> nil)
- iterate BasicMods 0..48 (#BasicMods is undefined with the 32/42/43
  holes and truncated the list)
- presets: fix the undefined `vocation` global in validation, the
  translateWheelVocation[...] table-index crash, a stray "[7]" literal,
  unify the equipedGems shape and pack export gems in domain order

+++ src/client/game.cpp
@@ -1533,7 +1533,7 @@ void Game::sendApplyWheelPoints(const std::vector<uint16_t>& slotPoints, uint16_
     m_protocolGame->sendApplyWheelPoints(slotPoints, greenGem, redGem, aquaGem, purpleGem);
 }
 
-void Game::gemAction(uint8_t actionType, uint8_t param, uint8_t pos)
+void Game::gemAction(uint8_t actionType, uint16_t param, uint8_t pos)
 {
     if (!canPerformGameAction())
         return;
